requirements.txt
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
joblib==1.3.2
scikit-learn==1.4.0
transformers==4.37.0
torch==2.2.0
pandas==2.2.0
numpy==1.26.3
pytest==7.4.4
pytest-asyncio==0.23.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
tenacity==8.2.3
pytz==2024.1
